refactor: extract chart component and helper function
